+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Add `minWidth` for phone number text input layout in `screen_registration_phone`
 - Add `Drug`s table
 - Migrate `BloodPressureHistoryScreen` to a fragment
+- Migrate `BloodSugarHistoryScreen` to a fragment
 
 ### Features
 - [In Progress: 06 Jul 2021] Add support for finding a patient online from ID scan